requirements.txt
# Основные библиотеки
python-telegram-bot==20.3
python-dotenv==1.0.0
